+++ src/components/facet/index.js
@@ -1,0 +1,7 @@
+export {
+  EuiFacetButton,
+} from './facet_button';
+
+export {
+  EuiFacetGroup,
+} from './facet_group';